Исправил ошибку с всплывающими подсказками при вводе неверных данных.
