+++ WPF_MailSender/MainWindow.xaml.cs
@@ -91,7 +91,6 @@ namespace WPF_MailSender
             }
         }
 
-
         #endregion
     }
 }
