+++ WPF_MailSender/Validation/RegExValidationRule.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Controls;
+
+namespace WPF_MailSender.Validation
+{
+    class RegExValidationRule : ValidationRule
+    {
+        public override ValidationResult Validate(object value, CultureInfo cultureInfo)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
